renamed DeleteTask file to DeleteTasks
